+++ package.json
@@ -61,7 +61,7 @@
       "fixpack"
     ]
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "prettier": {
     "singleQuote": true,
     "bracketSpacing": true,
