+++ test/sensitive-paths.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const HR = require('..');
+const HR = require('../src/index.js');
 
 test.beforeEach((t) => {
   const sensitivePaths = ['foo', 'bar', 'biz', 'arr', 'arr2', 'arr3'];
